saga/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
countries==0.2.0
Django==5.2a1
django-cors-headers==4.7.0
django-filter==25.1
django-htmx==1.21.0
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-ratelimit==4.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.1.0rc8
future==1.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.2
phonenumberslite==8.13.52
pillow==11.1.0
pluggy==1.5.0
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
setuptools==75.8.0
sqlparse==0.5.3
stripe==7.11.0
typing_extensions==4.13.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
